// models/teacher.model.ts
export class Teacher {
    id?:Number;
    username?: Number;
    password?: String;
    member_name?: String;
    member_alt_phone?: Number;
    email?: String;
    children?: Object[];
    schoolid?: Object;
    member_address?: String;
    role?:String;
    createdBy?:Number;
    modifiedBy?:Number;
    status?: String;
    data?: String;
}